revert(#664): moved api-structs back
Moved api-structs back from the lib into the binaries
again. Originally these were put into the lib as preparation
for a rust sdk-library, but instead it is not planned to
generate the lib from the openapi-specs. So for a cleaner
structure, they were moved back, where the are belong to.
