+++ src/binaries/hanami/src/api/http_endpoints/checkpoint/checkpoint_structs.rs
