+++ src/binaries/hanami/src/api/http_endpoints/checkpoint/get_checkpoint_v1_0.rs
@@ -19,8 +19,9 @@ use uuid::Uuid;
 
 use crate::database::checkpoint_table;
 
+use super::checkpoint_structs::CheckpointResp;
+
 use ainari_api::errors::ErrorResponse;
-use ainari_api::structs::checkpoint_structs::CheckpointResp;
 use ainari_api::user_context::UserContext;
 use ainari_common::enums;
