+++ src/binaries/hanami/src/api/http_endpoints/checkpoint/mod.rs
@@ -15,3 +15,4 @@
 pub mod delete_checkpoint_v1_0;
 pub mod get_checkpoint_v1_0;
 pub mod list_checkpoint_v1_0;
+pub mod checkpoint_structs;
