+++ src/binaries/hanami/src/api/http_endpoints/cluster/request_cluster_v1_0.rs
@@ -22,8 +22,9 @@ use validator::Validate;
 use crate::core::cluster_handler;
 use crate::database::cluster_table;
 
+use super::cluster_structs::{ClusterRequestReq, ClusterRequestResp};
+
 use ainari_api::errors::ErrorResponse;
-use ainari_api::structs::cluster_structs::{ClusterRequestReq, ClusterRequestResp};
 use ainari_api::user_context::UserContext;
 use ainari_common::enums;
 use ainari_common::error::AinariError;
