+++ src/binaries/hanami/src/api/http_endpoints/cluster/task/abort_task_v1_0.rs
@@ -21,8 +21,9 @@ use uuid::Uuid;
 use crate::database::cluster_table;
 use crate::database::task_table;
 
+use super::task_structs::{TaskResp, TaskState, TaskType};
+
 use ainari_api::errors::ErrorResponse;
-use ainari_api::structs::task_structs::{TaskResp, TaskState, TaskType};
 use ainari_api::user_context::UserContext;
 use ainari_common::enums;
