+++ src/binaries/hanami/src/api/http_endpoints/cluster/task/create_request_task_v1_0.rs
@@ -29,8 +29,9 @@ use crate::database::cluster_table;
 use crate::database::dataset_table;
 use crate::database::task_table;
 
+use super::task_structs::{TaskCreateRequestReq, TaskResp, TaskState, TaskType};
+
 use ainari_api::errors::ErrorResponse;
-use ainari_api::structs::task_structs::{TaskCreateRequestReq, TaskResp, TaskState, TaskType};
 use ainari_api::user_context::UserContext;
 use ainari_common::enums;
 use ainari_common::error::AinariError;
