+++ src/binaries/hanami/src/api/http_endpoints/cluster/task/create_train_task_v1_0.rs
@@ -27,8 +27,9 @@ use crate::database::cluster_table;
 use crate::database::dataset_table;
 use crate::database::task_table;
 
+use super::task_structs::{TaskCreateTrainReq, TaskResp, TaskState, TaskType};
+
 use ainari_api::errors::ErrorResponse;
-use ainari_api::structs::task_structs::{TaskCreateTrainReq, TaskResp, TaskState, TaskType};
 use ainari_api::user_context::UserContext;
 use ainari_common::enums;
 use ainari_dataset::dataset_io::read_data_set_file;
