+++ src/binaries/hanami/src/api/http_endpoints/cluster/task/mod.rs
@@ -19,3 +19,4 @@ pub mod create_request_task_v1_0;
 pub mod create_train_task_v1_0;
 pub mod get_task_v1_0;
 pub mod list_task_v1_0;
+pub mod task_structs;
