+++ src/binaries/hanami/src/api/http_endpoints/dataset/check_dataset_v1_0.rs
@@ -26,8 +26,9 @@ use validator::Validate;
 
 use crate::database::dataset_table;
 
+use super::dataset_structs::{DatasetCheckReq, DatasetCheckResp};
+
 use ainari_api::errors::ErrorResponse;
-use ainari_api::structs::dataset_structs::{DatasetCheckReq, DatasetCheckResp};
 use ainari_api::user_context::UserContext;
 use ainari_common::enums;
 use ainari_dataset::dataset_io::read_data_set_file;
