+++ src/binaries/hanami/src/api/http_endpoints/dataset/get_dataset_v1_0.rs
@@ -20,8 +20,9 @@ use uuid::Uuid;
 
 use crate::database::dataset_table;
 
+use super::dataset_structs::DatasetResp;
+
 use ainari_api::errors::ErrorResponse;
-use ainari_api::structs::dataset_structs::DatasetResp;
 use ainari_api::user_context::UserContext;
 use ainari_common::enums;
 use ainari_dataset::dataset_io::read_data_set_file;
