+++ src/binaries/torii/src/api/http_endpoints/auth/auth_structs.rs
