+++ src/binaries/torii/src/api/http_endpoints/auth/mod.rs
@@ -14,3 +14,4 @@
 
 pub mod create_token_v1_0;
 pub mod renew_token_v1_0;
+pub mod auth_structs;
