+++ src/binaries/torii/src/api/http_endpoints/auth/renew_token_v1_0.rs
@@ -17,8 +17,9 @@ use apistos::api_operation;
 
 use crate::api::token_handling;
 
+use super::auth_structs::UserTokenResp;
+
 use ainari_api::errors::ErrorResponse;
-use ainari_api::structs::auth_structs::UserTokenResp;
 use ainari_api::user_context::UserContext;
 
 #[api_operation(
