+++ src/binaries/torii/src/api/http_endpoints/project/create_project_v1_0.rs
@@ -19,8 +19,9 @@ use validator::Validate;
 
 use crate::database::project_table;
 
+use super::project_structs::{ProjectCreateReq, ProjectResp};
+
 use ainari_api::errors::ErrorResponse;
-use ainari_api::structs::project_structs::{ProjectCreateReq, ProjectResp};
 use ainari_api::user_context::UserContext;
 use ainari_common::enums;
