+++ src/binaries/torii/src/api/http_endpoints/project/get_project_v1_0.rs
@@ -18,8 +18,9 @@ use apistos::api_operation;
 
 use crate::database::project_table;
 
+use super::project_structs::ProjectResp;
+
 use ainari_api::errors::ErrorResponse;
-use ainari_api::structs::project_structs::ProjectResp;
 use ainari_api::user_context::UserContext;
 use ainari_common::enums;
