+++ src/binaries/torii/src/api/http_endpoints/user/mod.rs
@@ -16,3 +16,4 @@ pub mod create_user_v1_0;
 pub mod delete_user_v1_0;
 pub mod get_user_v1_0;
 pub mod list_user_v1_0;
+pub mod user_structs;
