+++ src/libs/rust/ainari_api/src/lib.rs
@@ -16,5 +16,4 @@
 
 pub mod endpoints;
 pub mod errors;
-pub mod structs;
 pub mod user_context;
